+++ breeze-dsl/breeze-dsl-bbcode/src/main/kotlin/BBCodeDsl.kt
@@ -1,0 +1,10 @@
+package com.windea.breezeframework.dsl.bbcode
+
+/**
+ * BBCode（Bulletin Board Code）的领域特定语言。
+ *
+ * 参见：[BBCode](https://www.bbcode.org/)
+ */
+@DslMarker
+@MustBeDocumented
+annotation class BBCodeDsl